requirements-py39.txt
# Core CLI and Config (Python 3.9 compatible)
click==8.1.7
pydantic==1.10.13
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
qdrant-client==1.7.0

# Machine Learning (Python 3.9 compatible)
numpy<2.0.0
torch==2.0.1
transformers==4.30.2
sentence-transformers==2.2.2

# HTTP and Data Processing
httpx==0.25.2
pandas==2.0.3
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
tenacity==8.2.3
tqdm==4.66.1
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development (Python 3.9 compatible)
ipython==8.12.3
black==23.12.1
